add base class for Friendly mobs
All friendly behaviors should inherit from this one
