+++ UnityProject/Assets/Resources/Prefabs/Mobs/NPC/Resources/Friendly/NPC_Ian.prefab
@@ -55,6 +55,17 @@ MonoBehaviour:
   isBleeding: 0
   livingHealthBehaviour: {fileID: 0}
   Severity: 0
+  armor:
+    Melee: 0
+    Bullet: 0
+    Laser: 0
+    Energy: 0
+    Bomb: 0
+    Rad: 0
+    Fire: 0
+    Acid: 0
+    Magic: 0
+    Bio: 0
 --- !u!1 &1088824078489892
 GameObject:
   m_ObjectHideFlags: 0
@@ -110,6 +121,17 @@ MonoBehaviour:
   isBleeding: 0
   livingHealthBehaviour: {fileID: 0}
   Severity: 0
+  armor:
+    Melee: 0
+    Bullet: 0
+    Laser: 0
+    Energy: 0
+    Bomb: 0
+    Rad: 0
+    Fire: 0
+    Acid: 0
+    Magic: 0
+    Bio: 0
 --- !u!1 &1178824245349430
 GameObject:
   m_ObjectHideFlags: 0
@@ -373,6 +395,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 9b91ecbcc199f4492b9a91e820070131, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  isDirty: 0
   sceneId: 0
   serverOnly: 0
   m_AssetId: 
@@ -489,6 +512,8 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   mobName: Ian
   knockedDownRotation: -180
+  minTimeBetweenRandomActions: 15
+  maxTimeBetweenRandomActions: 30
   allowedToGiveCommands: 0600000016000000
 --- !u!114 &1698261190327922250
 MonoBehaviour:
@@ -547,6 +572,7 @@ SpriteRenderer:
   m_MotionVectors: 1
   m_LightProbeUsage: 1
   m_ReflectionProbeUsage: 1
+  m_RayTracingMode: 0
   m_RenderingLayerMask: 4294967295
   m_RendererPriority: 0
   m_Materials:
