+++ UnityProject/Assets/Scripts/NPC/AI/Friendly.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 849b6b4aa04d40659000c4fa34e87b16
+timeCreated: 1587257809
