+++ UnityProject/Assets/Scripts/NPC/AI/Friendly/GenericFriendlyAI.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 3e28b1f5f04e4f699099349719baa668
+timeCreated: 1587266246
