+++ UnityProject/Assets/Scripts/NPC/AI/Friendly/MouseAI.cs.meta
