+++ UnityProject/Assets/Scripts/NPC/AI/Friendly/ParrotAI.cs.meta
