+++ UnityProject/Assets/Scripts/UI/Shuttles/GUI_ShuttleControl.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
+using NPC.AI.Friendly;
 using UnityEngine;
 
 /// Server only stuff
